// client/src/components/Ninja/ninja-text.enum.ts
export enum NinjaText {
  CONNECTING = 'CONNECTING',
  CONNECTED = 'CONNECTED',
  QR_CODE = 'QR_CODE',
  NO_NODES_FOUND = 'NO_NODES_FOUND',
  SHOW_MY_NODE = 'SHOW_MY_NODE',
  MY_NODE_COPIED = 'MY_NODE_COPIED',
  INVOICE_COPIED = 'INVOICE_COPIED',
  RECOMMENDATION_LIST = 'RECOMMENDATION_LIST',
  OPEN_CHANNEL = 'OPEN_CHANNEL',
  LOADING = 'LOADING',
}