// client/src/components/Ninja/tooltip.enum.ts
export enum TooltipKey {
  INVALID_PUB_KEY = 'INVALID_PUB_KEY',
  NINJA_HOVERED = 'NINJA_HOVERED',
  NINJA_CLICKED = 'NINJA_CLICKED',
  QR_CODE_CLICKED = 'QR_CODE_CLICKED',
  CONNECTIONS_HOVERED = 'CONNECTIONS_HOVERED',
  ADDRESS_CLICKED = 'ADDRESS_CLICKED',
  GRAPH_NOT_READY = 'GRAPH_NOT_READY',
  CHANNEL_OPENED = 'CHANNEL_OPENED',
  CHANNEL_OPENED_FAIL = 'CHANNEL_OPENED_FAIL',
}